+++ types/src/front/api_handlers/public/vaults.ts
@@ -2,11 +2,6 @@ import * as t from "io-ts";
 
 import { ContentNodeType } from "../../lib/connectors_api";
 
-export const ContentSchema = t.type({
-  dataSourceId: t.string,
-  parentsIn: t.array(t.string),
-});
-
 export const PostDataSourceViewSchema = t.type({
   dataSourceId: t.string,
   parentsIn: t.array(t.string),
@@ -31,41 +26,6 @@ export const PatchDataSourceViewSchema = t.union([
 export type PatchDataSourceViewType = t.TypeOf<
   typeof PatchDataSourceViewSchema
 >;
-
-const PostRestrictedVault = t.type({
-  memberIds: t.array(t.string),
-  isRestricted: t.literal(true),
-});
-
-const PostUnrestrictedVault = t.type({
-  memberIds: t.null,
-  isRestricted: t.literal(false),
-});
-
-export const PostVaultRequestBodySchema = t.intersection([
-  t.type({
-    name: t.string,
-  }),
-  t.union([PostRestrictedVault, PostUnrestrictedVault]),
-]);
-
-export type PostVaultRequestBodyType = t.TypeOf<
-  typeof PostVaultRequestBodySchema
->;
-
-export const PatchVaultRequestBodySchema = t.type({
-  name: t.union([t.string, t.undefined]),
-  content: t.union([t.array(ContentSchema), t.undefined]),
-});
-
-export type PatchVaultRequestBodyType = t.TypeOf<
-  typeof PatchVaultRequestBodySchema
->;
-
-export const PatchVaultMembersRequestBodySchema = t.union([
-  PostRestrictedVault,
-  PostUnrestrictedVault,
-]);
 
 export type LightContentNode = {
   dustDocumentId: string | null;
