+++ types/src/index.ts
@@ -17,6 +17,7 @@ export * from "./core/data_source";
 export * from "./front/acl";
 export * from "./front/api_handlers/internal/agent_configuration";
 export * from "./front/api_handlers/internal/assistant";
+export * from "./front/api_handlers/internal/vaults";
 export * from "./front/api_handlers/public/assistant";
 export * from "./front/api_handlers/public/data_sources";
 export * from "./front/api_handlers/public/vaults";
